+++ pages/site_sidebar.php
@@ -61,7 +61,7 @@
         </span>
       </a>
       <ul class="treeview-menu">
-        <!-- <li><a href="cdc_dashboard.php"><i class="fa fa-pie-chart"></i> Dashboard</a></li> -->
+        <li><a href="cdc_dashboard.php"><i class="fa fa-pie-chart"></i> Dashboard</a></li>
         <li><a href="cdc_organismo.php"><i class="fa fa-building"></i> Organismos</a></li>
         <li><a href="cdc_cliente.php"><i class="fa fa-user"></i> Clientes</a></li>
         <li class="treeview">